get v1 of dbcontext2 to actually work with sqlite queue
