+++ app/.gitignore
@@ -25,3 +25,5 @@ coverage
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+ios/DerivedData
