+++ app/src/contexts/DatabaseContext.tsx
@@ -176,321 +176,6 @@ export const DatabaseContextProvider: React.FC = (props) => {
   return <DatabaseContext.Provider value={databaseContext}>{props.children}</DatabaseContext.Provider>;
 };
 
-/* db query wrapper interface to hide db implementation */
-export enum QueryType {
-  DOC_TYPE_AND_ID = 'docType and ID',
-  DOC_TYPE = 'docType',
-  DOC_TYPE_AND_TRIP_ID = 'docType and trip_ID', //todo
-  DOC_TYPE_AND_BOUNDING_POLY = 'docType and bounding poly',
-  RAW_SQL = 'raw sql'
-}
-export interface IQuery {
-  type?: QueryType; // placeholder - might use to dictate what db
-  ID?: string;
-  docType?: DocType;
-  sql?: string;
-  boundingPoly?: GeoJSONObject;
-  geosOnly?: boolean;
-  limit?: number;
-  offset?: number;
-}
-
-export const query = async (queryConfig: IQuery, databaseContext: any) => {
-  /*  while (!databaseContext.sqliteDB) {
-    await setTimeout(() => {
-      alert('...waiting');
-    }, 3000);
-  }*/
-  try {
-    if (!databaseContext.sqliteDB.isAvailable || !databaseContext.db.isDBOpen()) {
-      alert('but unavailable');
-      return;
-    }
-  } catch (e) {
-    alert('crashing checking if available');
-    alert(JSON.stringify(e));
-  }
-  alert('supposedly available');
-  //alert(JSON.stringify(databaseContext));
-  if (Capacitor.getPlatform() != 'web') {
-    // alert('made it here');
-    console.log('typeof input');
-    console.log(typeof databaseContext);
-    let ret;
-    let db;
-    if (databaseContext.sqliteDB) {
-      db = databaseContext.sqliteDB;
-    } else {
-      return;
-    }
-    switch (queryConfig.type) {
-      case QueryType.DOC_TYPE_AND_ID:
-        ret = await db.query('select * from ' + queryConfig.docType + ' where id = ' + queryConfig.ID + ';\n');
-        if (!ret.values) {
-          //  db.close();
-          return false;
-        } else {
-          //  db.close();
-          return ret.values;
-        }
-      case QueryType.DOC_TYPE:
-        ret = await db.query(
-          'select * from ' + queryConfig.docType + (queryConfig.limit > 0 ? ' limit ' + queryConfig.limit + ';' : ';')
-        );
-        if (!ret.values) {
-          // db.close();
-          return false;
-        } else {
-          //db.close();
-          return ret.values;
-        }
-      case QueryType.RAW_SQL:
-        ret = await db.query(queryConfig.sql);
-        if (!ret.values) {
-          //db.close();
-          return false;
-        } else {
-          // db.close();
-          return ret.values;
-        }
-        break;
-      case QueryType.DOC_TYPE_AND_BOUNDING_POLY:
-        return await getByDocTypeAndBoudingPoly(queryConfig, db);
-      default:
-        alert(
-          'Your sqlite query needs a QueryType and corresponding parameters.  What you provided:  ' +
-            JSON.stringify(queryConfig)
-        );
-    }
-  }
-};
-
-const getByDocTypeAndBoudingPoly = async (queryConfig: IQuery, db: any) => {
-  return false;
-};
-
-/* db query wrapper interface to hide db implementation */
-export enum UpsertType {
-  DOC_TYPE_AND_ID = 'docType and ID',
-  DOC_TYPE_AND_ID_FAST_JSON_PATCH = 'docType and ID - FAST JSON PATCH', // uses sqlitejson1 extension when I get it working
-  DOC_TYPE_AND_ID_SLOW_JSON_PATCH = 'docType and ID - SLOW JSON PATCH', // workaround, all of these need to have same doctype
-  DOC_TYPE = 'docType',
-  RAW_SQL = 'raw sql'
-}
-export interface IUpsert {
-  type?: UpsertType; // placeholder - might use to dictate what db
-  ID?: string;
-  docType?: DocType;
-  sql?: string;
-  json?: Object;
-  geo?: GeoJSONObject; //todo - give geo it's own column so we can fetch many geos locally fast
-  trip_ID?: number; //todo - give a trip number for handling cache management
-}
-
-// v1: assumes all upsertconfigs are the same, will allow for multiple in v2
-export const upsert = async (upsertConfigs: Array<IUpsert>, databaseContext: any) => {
-  let batchUpdate = '';
-  let db;
-  if (databaseContext.sqliteDB) {
-    db = databaseContext.sqliteDB;
-  } else {
-    await setTimeout(() => {
-      console.log('waiting to upsert');
-    }, 1000);
-    // db = await openConnection(databaseContext);
-  }
-
-  let totalRecordsChanged = 0;
-
-  let ret: any;
-
-  // workaround until we get json1 extension working:
-  //TODO: change to return a count and share db like the other ones:
-  const patchSlowUpserts = await upsertConfigs.filter((e) => e.type == UpsertType.DOC_TYPE_AND_ID_SLOW_JSON_PATCH);
-  processSlowUpserts(patchSlowUpserts, databaseContext);
-
-  const docTypeAndIDUpserts = await upsertConfigs.filter((e) => e.type == UpsertType.DOC_TYPE_AND_ID);
-  if (docTypeAndIDUpserts.length >= 1) {
-    batchUpdate = buildSQLStringDOC_TYPE_AND_ID(docTypeAndIDUpserts);
-    ret = await handleExecute(batchUpdate, db);
-    if (ret != false) {
-      totalRecordsChanged += ret;
-    } else {
-      console.log('problem executing DOC_TYPE_AND_ID upserts:');
-      console.log('db failed upsert @' + totalRecordsChanged + ' records');
-    }
-  }
-
-  batchUpdate = '';
-  const everythingElse = await upsertConfigs.filter((e) => e.type != UpsertType.DOC_TYPE_AND_ID);
-  for (const upsertConfig of everythingElse) {
-    if (Capacitor.getPlatform() != 'web') {
-      // initialize the connection
-      // ret = db.execute(`select load_extension('json1');`); // not yet working
-
-      switch (upsertConfig.type) {
-        // full override update/upsert - json is replaced with new json
-        /*   case UpsertType.DOC_TYPE_AND_ID:
-          //the linter formatted this not me
-          batchUpdate +=
-            `insert into ` +
-            upsertConfig.docType +
-            ` (id,json) values ('` +
-            upsertConfig.ID +
-            `','` +
-            JSON.stringify(upsertConfig.json).split(`'`).join(`''`) +
-            //JSON.stringify(upsertConfig.json) +
-            `') on conflict(id) do update set json=excluded.json;\n`;
-          break;
-          */
-        // json patch upsert:
-        case UpsertType.DOC_TYPE_AND_ID_SLOW_JSON_PATCH:
-          break;
-        case UpsertType.DOC_TYPE_AND_ID_FAST_JSON_PATCH:
-          batchUpdate +=
-            `insert into ` +
-            upsertConfig.docType +
-            ` (id,json) values ('` +
-            upsertConfig.ID +
-            `','` +
-            JSON.stringify(upsertConfig.json).split(`'`).join(`''`) +
-            //JSON.stringify(upsertConfig.json) +
-            `') on conflict(id) do update set json_patch(json,excluded.json);\n`;
-          break;
-        // no ID present therefore these are inserts
-        case UpsertType.DOC_TYPE:
-          batchUpdate +=
-            `insert into ` +
-            upsertConfig.docType +
-            ` (json) values ('` +
-            JSON.stringify(upsertConfig.json).split(`'`).join(`''`) +
-            `')\n;`;
-          break;
-        // raw sql.
-        case UpsertType.RAW_SQL:
-          batchUpdate += upsertConfig.sql;
-          break;
-        default:
-          alert(
-            'Your sqlite query needs a UpsertType and corresponding parameters.  What you provided:  ' +
-              JSON.stringify(upsertConfig)
-          );
-      }
-    }
-  }
-  ret = await handleExecute(batchUpdate, db);
-  if (ret !== false) {
-    totalRecordsChanged += ret;
-  } else {
-    console.log('problem executing other upserts:');
-    console.log('db failed upsert @' + totalRecordsChanged + ' records');
-  }
-
-  return totalRecordsChanged;
-};
-
-const handleExecute = async (input: string, db: any) => {
-  let ret: any;
-  let batchUpdate = input;
-  if (batchUpdate !== '') {
-    ret = await db.execute(batchUpdate);
-    if (!ret.changes) {
-      //db.close();
-      return false;
-    } else {
-      //db.close();
-      return ret.changes.changes;
-    }
-  } else {
-    return 0;
-  }
-};
-//limit to all being the same type for now:
-const buildSQLStringDOC_TYPE_AND_ID = (upsertConfigs: Array<IUpsert>) => {
-  let batchUpdate = '';
-  if (upsertConfigs.length > 0) {
-    batchUpdate += `insert into ` + upsertConfigs[0].docType + ` (id,json) values `;
-    let rowCounter = 0;
-    for (const upsertConfig of upsertConfigs) {
-      batchUpdate += `('` + upsertConfig.ID + `','` + JSON.stringify(upsertConfig.json).split(`'`).join(`''`) + `')`;
-      if (rowCounter != upsertConfigs.length - 1) {
-        batchUpdate += ',';
-      }
-      rowCounter += 1;
-    }
-    batchUpdate += ` on conflict(id) do update set json=excluded.json;\n`;
-  }
-  return batchUpdate;
-};
-
-//limit to all being the same type for now:
-const processSlowUpserts = async (upsertConfigs: Array<IUpsert>, databaseContext: any) => {
-  if (upsertConfigs.length > 0) {
-    // get a list of the IDs and snag a doctype from one of them (THEY WILL ALL BE TREATED AS SAME TYPE)
-    let slowPatchOldIDS = [];
-    let slowPatchDocType;
-
-    for (const upsertConfig of upsertConfigs.filter((e) => {
-      return e.type == UpsertType.DOC_TYPE_AND_ID_SLOW_JSON_PATCH;
-    })) {
-      slowPatchOldIDS.push(upsertConfig.ID);
-      slowPatchDocType = upsertConfig.docType;
-    }
-
-    let ret;
-    //open db
-    let db;
-    if (databaseContext.sqliteDB) {
-      db = databaseContext.sqliteDB;
-    } else {
-      await setTimeout(() => {
-        console.log('waiting to slow upsert');
-      }, 1000);
-      //  db = await openConnection(databaseContext);
-    }
-
-    //build the query to get old records
-    const idsJSON = JSON.stringify(slowPatchOldIDS);
-    const idsString = idsJSON.substring(1, idsJSON.length - 1);
-    ret = await db.query('select * from ' + slowPatchDocType + ' where id in (' + idsString + ');');
-    const slowPatchOld = ret.values;
-
-    //patch them in memory
-    let batchUpdate = '';
-    console.log('*** building sql ***');
-
-    for (const upsertConfig of upsertConfigs) {
-      const old = JSON.parse(
-        slowPatchOld.filter((e) => {
-          return (e.id = upsertConfig.ID);
-        })[0].json
-      );
-      const patched =
-        `'` +
-        JSON.stringify({ ...old, ...upsertConfig.json })
-          .split(`'`)
-          .join(`''`) +
-        `'`;
-      batchUpdate +=
-        'insert into ' +
-        upsertConfig.docType +
-        ' (id, json) values (' +
-        upsertConfig.ID +
-        ',' +
-        patched +
-        ') on conflict (id) do update set json=excluded.json;';
-    }
-
-    //finally update them all:
-    console.log('*** executing sql ***');
-    if (batchUpdate != '') {
-      ret = db.execute(batchUpdate);
-    }
-    console.log('*** done sql ***');
-    //db.close();
-  }
-};
-
 export const createSqliteTables = async (databaseContext) => {
   // initialize the connection
   let db;
